@@ -1,12 +1,11 @@
 package com.gutotech.narutogame.helper;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.GestureDetector;
 import android.view.MotionEvent;
 import android.view.View;
 import android.widget.AdapterView;
-
 
 public class RecyclerItemClickListener implements RecyclerView.OnItemTouchListener {
 
