@@ -2,11 +2,7 @@ package com.gutotech.narutogame.model;
 
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.Exclude;
-import com.gutotech.narutogame.config.ConfigFirebase;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import com.gutotech.narutogame.config.FirebaseConfig;
 
 public class Player {
     private String id;
@@ -44,8 +40,8 @@ public class Player {
     }
 
     public void salvar() {
-        DatabaseReference reference = ConfigFirebase.getDatabase();
-        DatabaseReference player = reference.child("player").child(ConfigFirebase.getAuth().getCurrentUser().getUid());
+        DatabaseReference reference = FirebaseConfig.getDatabase();
+        DatabaseReference player = reference.child("player").child(FirebaseConfig.getAuth().getCurrentUser().getUid());
         player.setValue(this);
     }
 
